test: add pfm upgrade test too
