+++ app/test/upgrade_test.go
@@ -1,9 +1,12 @@
 package app_test
 
 import (
-	"testing"
-
+	"encoding/json"
 	"fmt"
+	"strings"
+	"testing"
+	"time"
+
 	app "github.com/celestiaorg/celestia-app/v2/app"
 	"github.com/celestiaorg/celestia-app/v2/app/encoding"
 	v1 "github.com/celestiaorg/celestia-app/v2/pkg/appconsts/v1"
@@ -19,11 +22,9 @@ import (
 	tmproto "github.com/tendermint/tendermint/proto/tendermint/types"
 	tmversion "github.com/tendermint/tendermint/proto/tendermint/version"
 	dbm "github.com/tendermint/tm-db"
-	"encoding/json"
 
+	packetforwardtypes "github.com/cosmos/ibc-apps/middleware/packet-forward-middleware/v6/router/types"
 	"github.com/tendermint/tendermint/libs/log"
-	"strings"
-	"time"
 )
 
 func TestUpgradeAppVersion(t *testing.T) {
@@ -46,6 +47,8 @@ func TestUpgradeAppVersion(t *testing.T) {
 	require.EqualValues(t, 2, testApp.AppVersion())
 }
 
+// TestMinFeeDuringVersionUpgrades verifies that the minfee module's params are overridden during an
+// upgrade from v1 -> v2.
 func TestMinFeeDuringVersionUpgrades(t *testing.T) {
 	testApp, _ := SetupTestAppWithUpgradeHeight(t, 3)
 
@@ -101,6 +104,8 @@ func TestMinFeeDuringVersionUpgrades(t *testing.T) {
 	require.Equal(t, want.String(), strings.Trim(got.Param.Value, "\""))
 }
 
+// TestICADuringVersionUpgrades verifies that the ICA module's params are overridden during an
+// upgrade from v1 -> v2.
 func TestICADuringVersionUpgrades(t *testing.T) {
 	testApp, _ := SetupTestAppWithUpgradeHeight(t, 3)
 	ctx := testApp.NewContext(true, tmproto.Header{
@@ -134,6 +139,52 @@ func TestICADuringVersionUpgrades(t *testing.T) {
 	})
 	require.NoError(t, err)
 	require.Equal(t, "true", got.Param.Value)
+}
+
+// TestPacketForwardMiddlewareAgainstAppUpgrades verifies that the PFM module's params are overridden during an
+// upgrade from v1 -> v2.
+func TestPacketForwardMiddlewareAgainstAppUpgrades(t *testing.T) {
+	testApp, _ := SetupTestAppWithUpgradeHeight(t, 3)
+	supportedVersions := []uint64{v1.Version, v2.Version}
+	require.Equal(t, supportedVersions, testApp.SupportedVersions())
+
+	ctx := testApp.NewContext(true, tmproto.Header{
+		Version: tmversion.Consensus{
+			App: 1,
+		},
+	})
+	testApp.BeginBlock(abci.RequestBeginBlock{Header: tmproto.Header{
+		Height:  2,
+		Version: tmversion.Consensus{App: 1},
+	}})
+	// app version should not have changed yet
+	require.EqualValues(t, 1, testApp.AppVersion())
+	// PacketForwardMiddleware should not have been set yet
+	gotBefore, err := testApp.ParamsKeeper.Params(ctx, &proposal.QueryParamsRequest{
+		Subspace: packetforwardtypes.ModuleName,
+		Key:      string(packetforwardtypes.KeyFeePercentage),
+	})
+	require.Equal(t, "", gotBefore.Param.Value)
+	require.NoError(t, err)
+	// now the app version changes
+	respEndBlock := testApp.EndBlock(abci.RequestEndBlock{Height: 2})
+	testApp.Commit()
+	require.NotNil(t, respEndBlock.ConsensusParamUpdates.Version)
+	require.EqualValues(t, 2, respEndBlock.ConsensusParamUpdates.Version.AppVersion)
+	require.EqualValues(t, 2, testApp.AppVersion())
+	// create a new context after endBlock
+	newCtx := testApp.NewContext(true, tmproto.Header{
+		Version: tmversion.Consensus{
+			App: 2,
+		},
+	})
+	got, err := testApp.ParamsKeeper.Params(newCtx, &proposal.QueryParamsRequest{
+		Subspace: packetforwardtypes.ModuleName,
+		Key:      string(packetforwardtypes.KeyFeePercentage),
+	})
+	require.NoError(t, err)
+	require.NoError(t, err)
+	require.Equal(t, "0.000000000000000000", strings.Trim(got.Param.Value, "\""))
 }
 
 func SetupTestAppWithUpgradeHeight(t *testing.T, upgradeHeight int64) (*app.App, keyring.Keyring) {
